Fixed bug where gzip content would create nuts progress reports.
This fix is an ugly work around. The player will not give access to the
response headers, so we have to guees. When gzip compression is turned
on you will end up with changing, unreliable, smaller than bytesLoaded
values for bytesTotal. So what we do (not sure if it is **the** best
answer here) is to check for moot values and make sure that:
  - bytesRemaining will always be 0 when loading is done.
  - bytesTotal will always be equal to bytesLoaded when loading is done.
  - Give no bizarre percentage midway